fix: address Tavily review findings
