+++ src/tools/search/search.ts
@@ -2,6 +2,7 @@ import axios from 'axios';
 import { RecursiveCharacterTextSplitter } from '@langchain/textsplitters';
 import type * as t from './types';
 import { getAttribution, createDefaultLogger } from './utils';
+import { createTavilyAPI } from './tavily-search';
 import { BaseReranker } from './rerankers';
 
 const chunker = {
@@ -408,415 +409,6 @@ const createSearXNGAPI = (
   return { getSources };
 };
 
-const TAVILY_TIME_RANGE_MAP: Record<string, string> = {
-  h: 'day',
-  d: 'day',
-  w: 'week',
-  m: 'month',
-  y: 'year',
-} as const;
-
-const DEFAULT_TAVILY_TIMEOUT = 15000;
-
-const TAVILY_COUNTRY_ALIASES: Record<string, string> = {
-  ba: 'bosnia and herzegovina',
-  cg: 'congo',
-  cz: 'czech republic',
-  gb: 'united kingdom',
-  mm: 'myanmar',
-  tr: 'turkey',
-  tt: 'trinidad and tobago',
-  uk: 'united kingdom',
-};
-
-const TAVILY_SUPPORTED_COUNTRIES = new Set([
-  'afghanistan',
-  'albania',
-  'algeria',
-  'andorra',
-  'angola',
-  'argentina',
-  'armenia',
-  'australia',
-  'austria',
-  'azerbaijan',
-  'bahamas',
-  'bahrain',
-  'bangladesh',
-  'barbados',
-  'belarus',
-  'belgium',
-  'belize',
-  'benin',
-  'bhutan',
-  'bolivia',
-  'bosnia and herzegovina',
-  'botswana',
-  'brazil',
-  'brunei',
-  'bulgaria',
-  'burkina faso',
-  'burundi',
-  'cambodia',
-  'cameroon',
-  'canada',
-  'cape verde',
-  'central african republic',
-  'chad',
-  'chile',
-  'china',
-  'colombia',
-  'comoros',
-  'congo',
-  'costa rica',
-  'croatia',
-  'cuba',
-  'cyprus',
-  'czech republic',
-  'denmark',
-  'djibouti',
-  'dominican republic',
-  'ecuador',
-  'egypt',
-  'el salvador',
-  'equatorial guinea',
-  'eritrea',
-  'estonia',
-  'ethiopia',
-  'fiji',
-  'finland',
-  'france',
-  'gabon',
-  'gambia',
-  'georgia',
-  'germany',
-  'ghana',
-  'greece',
-  'guatemala',
-  'guinea',
-  'haiti',
-  'honduras',
-  'hungary',
-  'iceland',
-  'india',
-  'indonesia',
-  'iran',
-  'iraq',
-  'ireland',
-  'israel',
-  'italy',
-  'jamaica',
-  'japan',
-  'jordan',
-  'kazakhstan',
-  'kenya',
-  'kuwait',
-  'kyrgyzstan',
-  'latvia',
-  'lebanon',
-  'lesotho',
-  'liberia',
-  'libya',
-  'liechtenstein',
-  'lithuania',
-  'luxembourg',
-  'madagascar',
-  'malawi',
-  'malaysia',
-  'maldives',
-  'mali',
-  'malta',
-  'mauritania',
-  'mauritius',
-  'mexico',
-  'moldova',
-  'monaco',
-  'mongolia',
-  'montenegro',
-  'morocco',
-  'mozambique',
-  'myanmar',
-  'namibia',
-  'nepal',
-  'netherlands',
-  'new zealand',
-  'nicaragua',
-  'niger',
-  'nigeria',
-  'north korea',
-  'north macedonia',
-  'norway',
-  'oman',
-  'pakistan',
-  'panama',
-  'papua new guinea',
-  'paraguay',
-  'peru',
-  'philippines',
-  'poland',
-  'portugal',
-  'qatar',
-  'romania',
-  'russia',
-  'rwanda',
-  'saudi arabia',
-  'senegal',
-  'serbia',
-  'singapore',
-  'slovakia',
-  'slovenia',
-  'somalia',
-  'south africa',
-  'south korea',
-  'south sudan',
-  'spain',
-  'sri lanka',
-  'sudan',
-  'sweden',
-  'switzerland',
-  'syria',
-  'taiwan',
-  'tajikistan',
-  'tanzania',
-  'thailand',
-  'togo',
-  'trinidad and tobago',
-  'tunisia',
-  'turkey',
-  'turkmenistan',
-  'uganda',
-  'ukraine',
-  'united arab emirates',
-  'united kingdom',
-  'united states',
-  'uruguay',
-  'uzbekistan',
-  'venezuela',
-  'vietnam',
-  'yemen',
-  'zambia',
-  'zimbabwe',
-]);
-
-const TAVILY_REGION_NAMES = new Intl.DisplayNames(['en'], {
-  type: 'region',
-});
-
-const normalizeTavilyCountryName = (country: string): string =>
-  country
-    .toLowerCase()
-    .replace(/\s*\([^)]*\)/g, '')
-    .replace(/\s*&\s*/g, ' and ')
-    .normalize('NFD')
-    .replace(/\p{Diacritic}/gu, '');
-
-const normalizeTavilyCountry = (country?: string): string | undefined => {
-  const normalizedCountry = country?.trim().toLowerCase();
-  if (normalizedCountry == null || normalizedCountry === '') {
-    return undefined;
-  }
-
-  const countryAlias = TAVILY_COUNTRY_ALIASES[normalizedCountry];
-  if (countryAlias != null) {
-    return TAVILY_SUPPORTED_COUNTRIES.has(countryAlias)
-      ? countryAlias
-      : undefined;
-  }
-
-  if (/^[a-z]{2}$/.test(normalizedCountry)) {
-    const regionName = TAVILY_REGION_NAMES.of(normalizedCountry.toUpperCase());
-    if (regionName == null) {
-      return undefined;
-    }
-    const countryName = normalizeTavilyCountryName(regionName);
-    return TAVILY_SUPPORTED_COUNTRIES.has(countryName)
-      ? countryName
-      : undefined;
-  }
-
-  const countryName = normalizeTavilyCountryName(normalizedCountry);
-  return TAVILY_SUPPORTED_COUNTRIES.has(countryName) ? countryName : undefined;
-};
-
-const getHostname = (link: string): string => {
-  try {
-    return new URL(link).hostname;
-  } catch {
-    return link;
-  }
-};
-
-const createTavilyAPI = (
-  apiKey?: string,
-  apiUrl?: string,
-  options?: t.TavilySearchOptions
-): {
-  getSources: (params: t.GetSourcesParams) => Promise<t.SearchResult>;
-} => {
-  const config = {
-    apiKey: apiKey ?? process.env.TAVILY_API_KEY,
-    apiUrl:
-      apiUrl ??
-      process.env.TAVILY_SEARCH_URL ??
-      'https://api.tavily.com/search',
-    timeout: options?.timeout ?? DEFAULT_TAVILY_TIMEOUT,
-  };
-
-  if (config.apiKey == null || config.apiKey === '') {
-    throw new Error('TAVILY_API_KEY is required for Tavily API');
-  }
-
-  const getSources = async ({
-    query,
-    date,
-    country,
-    numResults = 8,
-    type,
-    news,
-  }: t.GetSourcesParams): Promise<t.SearchResult> => {
-    if (!query.trim()) {
-      return { success: false, error: 'Query cannot be empty' };
-    }
-
-    try {
-      const timeRange =
-        options?.timeRange ??
-        (date != null ? (TAVILY_TIME_RANGE_MAP[date] ?? 'day') : undefined);
-      const topic =
-        options?.topic ??
-        (news === true || type === 'news' ? 'news' : 'general');
-      const maxResults = options?.maxResults ?? numResults;
-      const searchDepth = options?.searchDepth ?? 'basic';
-
-      const payload: Record<string, unknown> = {
-        query,
-        search_depth: searchDepth,
-        topic,
-        max_results: Math.min(Math.max(1, maxResults), 20),
-      };
-
-      if (
-        options?.safeSearch != null &&
-        searchDepth !== 'fast' &&
-        searchDepth !== 'ultra-fast'
-      ) {
-        payload.safe_search = options.safeSearch;
-      }
-      if (timeRange != null) {
-        payload.time_range = timeRange;
-      }
-      const tavilyCountry =
-        topic === 'general' ? normalizeTavilyCountry(country) : undefined;
-      if (tavilyCountry != null) {
-        payload.country = tavilyCountry;
-      }
-      if (type === 'images' || options?.includeImages) {
-        payload.include_images = true;
-      }
-      if (options?.includeAnswer != null) {
-        payload.include_answer = options.includeAnswer;
-      }
-      if (options?.includeRawContent != null) {
-        payload.include_raw_content = options.includeRawContent;
-      }
-      if (
-        options?.includeDomains != null &&
-        options.includeDomains.length > 0
-      ) {
-        payload.include_domains = options.includeDomains;
-      }
-      if (
-        options?.excludeDomains != null &&
-        options.excludeDomains.length > 0
-      ) {
-        payload.exclude_domains = options.excludeDomains;
-      }
-      if (options?.includeImageDescriptions != null) {
-        payload.include_image_descriptions = options.includeImageDescriptions;
-      }
-      if (options?.includeFavicon != null) {
-        payload.include_favicon = options.includeFavicon;
-      }
-      if (options?.chunksPerSource != null && searchDepth === 'advanced') {
-        payload.chunks_per_source = options.chunksPerSource;
-      }
-
-      const response = await axios.post<t.TavilySearchResponse>(
-        config.apiUrl,
-        payload,
-        {
-          headers: {
-            Authorization: `Bearer ${config.apiKey}`,
-            'Content-Type': 'application/json',
-          },
-          timeout: config.timeout,
-        }
-      );
-
-      const data = response.data;
-
-      const organicResults: t.OrganicResult[] = (data.results ?? []).map(
-        (result: t.TavilySearchResult) => ({
-          title: result.title ?? '',
-          link: result.url ?? '',
-          snippet: result.content ?? '',
-          date: result.published_date,
-        })
-      );
-
-      const imageResults: t.ImageResult[] = Array.isArray(data.images)
-        ? data.images
-            .slice(0, 6)
-            .reduce<t.ImageResult[]>((acc, image, index) => {
-              const imageUrl = typeof image === 'string' ? image : image.url;
-              if (imageUrl == null || imageUrl === '') {
-                return acc;
-              }
-              acc.push({
-                imageUrl,
-                title:
-                  typeof image === 'string' ? undefined : image.description,
-                position: index + 1,
-              });
-              return acc;
-            }, [])
-        : [];
-
-      const newsResults: t.NewsResult[] =
-        topic === 'news'
-          ? organicResults.map((r) => ({
-              title: r.title,
-              link: r.link,
-              snippet: r.snippet,
-              date: r.date,
-              source: getHostname(r.link),
-            }))
-          : [];
-
-      const results: t.SearchResultData = {
-        organic: organicResults,
-        images: imageResults,
-        topStories: [],
-        videos: [],
-        news: newsResults,
-        answerBox: data.answer != null ? { snippet: data.answer } : undefined,
-        relatedSearches: [],
-      };
-
-      return { success: true, data: results };
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : String(error);
-      return {
-        success: false,
-        error: `Tavily API request failed: ${errorMessage}`,
-      };
-    }
-  };
-
-  return { getSources };
-};
-
 export const createSearchAPI = (
   config: t.SearchConfig
 ): {
@@ -874,9 +466,7 @@ export const createSourceProcessor = (
   ): t.ScrapeResult => {
     const rawMetadata = scraper.extractMetadata(response);
     const metadata =
-      Object.keys(rawMetadata).length > 0
-        ? (rawMetadata as t.ScrapeMetadata)
-        : undefined;
+      Object.keys(rawMetadata).length > 0 ? rawMetadata : undefined;
     const attribution = getAttribution(url, metadata, logger_);
 
     if (response.success && response.data) {
@@ -949,11 +539,14 @@ export const createSourceProcessor = (
           );
         }
 
-        const results = responses.map(([url, response]) =>
-          processResponse(url, response)
-        );
         const withHighlights = await Promise.all(
-          results.map((result) => addHighlights(result, query, onGetHighlights))
+          responses.map(([url, response]) =>
+            addHighlights(
+              processResponse(url, response),
+              query,
+              onGetHighlights
+            )
+          )
         );
         return withHighlights;
       } catch (error) {
